Mess around with using react context to refactor config (WIP)
